requirements.txt
# Dependencias del Sistema de Pesaje de Fardos
# Instalar con: pip install -r requirements.txt

# Biblioteca para generar PDFs
reportlab>=3.6.0

# Biblioteca para comunicación serie con balanzas
pyserial>=3.5

# Bibliotecas estándar (ya incluidas con Python)
# tkinter - Interfaz gráfica (incluida con Python)
# csv - Manejo de archivos CSV (incluida con Python)
# datetime - Manejo de fechas (incluida con Python)
# os - Operaciones del sistema (incluida con Python)
# random - Números aleatorios (incluida con Python)
# time - Funciones de tiempo (incluida con Python)
# typing - Anotaciones de tipos (incluida con Python 3.5+)

# Dependencias opcionales para funcionalidades avanzadas
# pandas>=1.3.0  # Para análisis de datos avanzado
# matplotlib>=3.5.0  # Para gráficos y visualizaciones
# openpyxl>=3.0.0  # Para exportar a Excel
